AI/pdf_qa/requirements.txt
langchain
langchain-community
openai
fastapi
uvicorn
python-dotenv
pymupdf
faiss-cpu
tiktoken